# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LIFELINK emergency response, dispatch, routing, hospital readiness, and demo simulation API.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /emergencies:
    get:
      operationId: listEmergencies
      tags: [emergencies]
      summary: List active and recent emergencies
      responses:
        "200":
          description: Emergency list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Emergency"
    post:
      operationId: createEmergency
      tags: [emergencies]
      summary: Create an SOS emergency and dispatch a suitable ambulance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmergencyInput"
      responses:
        "201":
          description: Emergency created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Emergency"
        "400":
          description: Invalid emergency
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /emergencies/{id}:
    get:
      operationId: getEmergency
      tags: [emergencies]
      summary: Get emergency details
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Emergency detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Emergency"
        "404":
          description: Emergency not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /emergencies/{id}/decision:
    get:
      operationId: getEmergencyDecision
      tags: [routing]
      summary: Get the current route and hospital recommendation
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Decision explanation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Decision"
        "404":
          description: Decision not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /ambulances/{id}/accept:
    post:
      operationId: acceptEmergency
      tags: [ambulances]
      summary: Accept an emergency dispatch
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AmbulanceEmergencyInput"
      responses:
        "200":
          description: Dispatch accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ambulance"
  /ambulances/{id}/decline:
    post:
      operationId: declineEmergency
      tags: [ambulances]
      summary: Decline an emergency dispatch
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AmbulanceEmergencyInput"
      responses:
        "200":
          description: Dispatch declined and reassigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Emergency"
  /ambulances/{id}/status:
    post:
      operationId: updateAmbulanceStatus
      tags: [ambulances]
      summary: Update ambulance service status
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AmbulanceStatusInput"
      responses:
        "200":
          description: Updated ambulance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ambulance"
  /ambulances/{id}/location:
    post:
      operationId: updateAmbulanceLocation
      tags: [ambulances]
      summary: Send an ambulance GPS update
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationInput"
      responses:
        "200":
          description: Updated ambulance location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ambulance"
  /ambulances/{id}/arrived:
    post:
      operationId: markArrivedAtScene
      tags: [ambulances]
      summary: Mark arrival at the emergency scene
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AmbulanceEmergencyInput"
      responses:
        "200":
          description: Updated ambulance and emergency
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Emergency"
  /ambulances/{id}/patient-onboard:
    post:
      operationId: markPatientOnboard
      tags: [ambulances]
      summary: Mark patient onboard and begin hospital transport
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AmbulanceEmergencyInput"
      responses:
        "200":
          description: Updated emergency
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Emergency"
  /ambulances/{id}/complete:
    post:
      operationId: completeEmergency
      tags: [ambulances]
      summary: Complete a hospital handoff
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AmbulanceEmergencyInput"
      responses:
        "200":
          description: Completed emergency
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Emergency"
...
  /hospitals/{id}/status:
    post:
      operationId: updateHospitalStatus
      tags: [hospitals]
      summary: Update hospital readiness and capacity
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HospitalStatusInput"
      responses:
        "200":
          description: Updated hospital
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hospital"
  /routes/calculate:
    post:
      operationId: calculateRoutes
      tags: [routing]
      summary: Calculate traffic-aware route alternatives
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteRequest"
      responses:
        "200":
          description: Ranked route alternatives
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RouteOption"
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Emergency:
      type: object
      properties:
        id: { type: integer }
        patientName: { type: string }
        emergencyType: { type: string }
        severity: { type: string, enum: [CRITICAL, HIGH, MEDIUM, LOW] }
        status: { type: string, enum: [NEW, DISPATCHED, ACCEPTED, EN_ROUTE, ON_SCENE, TRANSPORTING, COMPLETED, CANCELLED] }
        latitude: { type: number }
        longitude: { type: number }
        locationLabel: { type: string }
        requiredCapabilities:
          type: array
          items: { type: string }
        assignedAmbulanceId: { type: ["integer", "null"] }
        selectedHospitalId: { type: ["integer", "null"] }
        etaMinutes: { type: ["integer", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, patientName, emergencyType, severity, status, latitude, longitude, locationLabel, requiredCapabilities, assignedAmbulanceId, selectedHospitalId, etaMinutes, createdAt, updatedAt]
    EmergencyInput:
      type: object
      properties:
        patientName: { type: string, minLength: 1, maxLength: 100 }
        emergencyType: { type: string, minLength: 1, maxLength: 80 }
        severity: { type: string, enum: [CRITICAL, HIGH, MEDIUM, LOW] }
        latitude: { type: number, minimum: -90, maximum: 90 }
        longitude: { type: number, minimum: -180, maximum: 180 }
        locationLabel: { type: string, minLength: 1, maxLength: 200 }
        requiredCapabilities:
          type: array
          items: { type: string }
      required: [patientName, emergencyType, severity, latitude, longitude, locationLabel, requiredCapabilities]
    Ambulance:
      type: object
      properties:
        id: { type: integer }
        callSign: { type: string }
        vehicleNumber: { type: string }
        driverName: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        status: { type: string, enum: [AVAILABLE, EN_ROUTE, ON_SCENE, TRANSPORTING, OFFLINE] }
        equipment:
          type: array
          items: { type: string }
        currentEmergencyId: { type: ["integer", "null"] }
        destinationHospitalId: { type: ["integer", "null"] }
        etaMinutes: { type: ["integer", "null"] }
        lastUpdated: { type: string, format: date-time }
      required: [id, callSign, vehicleNumber, driverName, latitude, longitude, status, equipment, currentEmergencyId, destinationHospitalId, etaMinutes, lastUpdated]
    AmbulanceEmergencyInput:
      type: object
      properties:
        emergencyId: { type: integer }
      required: [emergencyId]
    AmbulanceStatusInput:
      type: object
      properties:
        status: { type: string, enum: [AVAILABLE, EN_ROUTE, ON_SCENE, TRANSPORTING, OFFLINE] }
      required: [status]
    LocationInput:
      type: object
      properties:
        latitude: { type: number, minimum: -90, maximum: 90 }
        longitude: { type: number, minimum: -180, maximum: 180 }
        speedKph: { type: number, minimum: 0, maximum: 250 }
        heading: { type: number, minimum: 0, maximum: 360 }
      required: [latitude, longitude, speedKph, heading]
...
    HospitalStatusInput:
      type: object
      properties:
        readinessStatus: { type: string, enum: [READY, BUSY, FULL] }
        emergencyStatus: { type: string, enum: [ACCEPTING, LIMITED, DIVERTING] }
        traumaBeds: { type: integer, minimum: 0 }
        icuBeds: { type: integer, minimum: 0 }
        ventilators: { type: integer, minimum: 0 }
        waitMinutes: { type: integer, minimum: 0 }
      required: [readinessStatus, emergencyStatus, traumaBeds, icuBeds, ventilators, waitMinutes]
    RouteRequest:
      type: object
      properties:
        emergencyId: { type: integer }
      required: [emergencyId]
...
    Decision:
      type: object
      properties:
        emergencyId: { type: integer }
        selectedRoute: { $ref: "#/components/schemas/RouteOption" }
        selectedHospitalId: { type: ["integer", "null"] }
        selectedHospitalName: { type: ["string", "null"] }
        confidence: { type: integer }
        reasons:
          type: array
          items: { type: string }
        warnings:
          type: array
          items: { type: string }
        updatedAt: { type: string, format: date-time }
      required: [emergencyId, selectedRoute, selectedHospitalId, selectedHospitalName, confidence, reasons, warnings, updatedAt]
...