# lib/api-zod/src/generated/types/routeRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LIFELINK emergency response, dispatch, routing, hospital readiness, and demo simulation API.
 * OpenAPI spec version: 0.1.0
 */

export interface RouteRequest {
  emergencyId: number;
}